refactor: LottoMachine을 제거한 뒤 FixedLottosGenerator,RandomLottosGenerator로 대체
