--- src/main/kotlin/lotto/model/LottoMachine.kt
@@ -1,10 +1,0 @@
-package lotto.model
-
-object LottoMachine {
-    fun createLottos(
-        numberOfLottos: Int,
-        lottosGenerator: LottosGenerator,
-    ): List<Lotto> {
-        return lottosGenerator.generate(numberOfLottos)
-    }
-}
